fix(remediate): degraded gather evidence falls the matrix back to the static task list (review of #342)
A fail-open gather (an unreadable baseline, no floor evidence anywhere)
yields a zero-order plan that proves nothing; deriving an empty matrix
from it silently turned the whole schedule off every week. The gather
now reports structural evidence degradation, and the matrix derivation
treats it exactly like the planning-throw path: static policy task
list, with the reason disclosed in the plan output and the workflow
notices. Healthy evidence with zero orders remains the legitimate
no-job firing. The deferral join moves to its own module (deferrals.ts,
the module-size split the self-guardrail's large-file gate asked for).
